pyproject.toml
[project]
name = "teen-mh"
version = "0.1.0"
description = "End-to-end ML pipeline for teen mental-health (depression) classification"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "DeryFerd" }]
license = { text = "MIT" }
dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "scikit-learn>=1.5",
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "pyyaml>=6.0",
    "joblib>=1.4",
    "mlflow>=2.16",
    "imbalanced-learn>=0.12",
    "xgboost>=2.1",
    "lightgbm>=4.5",
    "optuna>=4.0",
    "pydantic>=2.9",
    "rich>=13.9",
    "typer>=0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "pre-commit>=3.8",
]

[project.scripts]
teen-mh = "teen_mh.cli:app"

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/teen_mh"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
pythonpath = ["src"]
